requirements.txt
Django==4.2.9
psycopg2-binary==2.9.9
gunicorn==21.2.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
Pillow==10.2.0
python-decouple==3.8
